+++ packages/jaspr/lib/src/framework/framework.dart
@@ -12,25 +12,24 @@ import 'package:meta/meta.dart';
 
 import '../foundation/basic_types.dart';
 import '../foundation/binding.dart';
-import '../foundation/scheduler.dart';
 import '../foundation/sync.dart';
 import '../ui/styles/styles.dart';
 
 part 'build_context.dart';
 part 'build_owner.dart';
 part 'components_binding.dart';
-part 'render_element.dart';
-part 'render_scope.dart';
 part 'dom_component.dart';
 part 'inactive_elements.dart';
 part 'inherited_component.dart';
 part 'keys.dart';
 part 'multi_child_element.dart';
+part 'observer_component.dart';
+part 'render_element.dart';
+part 'render_scope.dart';
 part 'single_child_element.dart';
 part 'state_mixins.dart';
 part 'stateful_component.dart';
 part 'stateless_component.dart';
-part 'observer_component.dart';
 
 /// Describes the configuration for an [Element].
 ///
@@ -222,29 +221,34 @@ abstract class Element implements BuildContext {
   Component? _component;
 
   /// The root component binding that manages the component tree.
-  BuildOwner? _owner;
-  BuildOwner get owner => _owner!;
+  @override
+  AppBinding get binding => _binding!;
+  AppBinding? _binding;
 
-  Renderer? _renderer;
+  /// The root build owner that manages the build cycle.
+  BuildOwner get owner => _owner!;
+  BuildOwner? _owner;
+
   Renderer get renderer => _renderer!;
+  Renderer? _renderer;
 
   /// The previous sibling element.
   Element? _prevSibling;
 
   /// The previous ancestor sibling
-  Element? _prevAncestorSibling;
   Element? get prevAncestorSibling => _prevAncestorSibling;
+  Element? _prevAncestorSibling;
 
   /// The last child element.
   Element? _lastChild;
 
   /// The last child dom node.
-  RenderElement? _lastNode;
   RenderElement? get lastNode => _lastNode;
+  RenderElement? _lastNode;
 
   /// The nearest ancestor dom node.
-  RenderElement? _parentNode;
   RenderElement? get parentNode => _parentNode;
+  RenderElement? _parentNode;
 
   // This is used to verify that Element objects move through life in an
   // orderly fashion.
@@ -378,13 +382,15 @@ abstract class Element implements BuildContext {
 
     if (parent != null) {
       _owner = parent.owner;
+      _binding = parent.binding;
       _renderer = _inheritRendererFromParent();
     }
     assert(_owner != null);
+    assert(_binding != null);
 
     final Key? key = component.key;
     if (key is GlobalKey) {
-      ComponentsBinding.instance!._registerGlobalKey(key, this);
+      ComponentsBinding._registerGlobalKey(key, this);
     }
     _updateInheritance();
     _updateObservers();
@@ -569,6 +575,7 @@ abstract class Element implements BuildContext {
     assert(_lifecycleState == _ElementLifecycle.inactive);
     assert(_component != null);
     assert(_owner != null);
+    assert(_binding != null);
     assert(_parent != null);
     assert(_depth != null);
     final bool hadDependencies = (_dependencies != null && _dependencies!.isNotEmpty) || _hadUnsatisfiedDependencies;
@@ -649,7 +656,7 @@ abstract class Element implements BuildContext {
 
     final Key? key = component.key;
     if (key is GlobalKey) {
-      ComponentsBinding.instance!._unregisterGlobalKey(key, this);
+      ComponentsBinding._unregisterGlobalKey(key, this);
     }
 
     _parentNode = null;