+++ packages/jaspr/lib/src/server/native_interop.dart
@@ -1,5 +1,5 @@
 import '../framework/framework.dart';
 
 extension NativeDomNode on RenderElement {
-  dynamic get nativeElement => getData();
+  dynamic get nativeElement => renderData;
 }