+++ packages/jaspr/lib/src/server/server_handler.dart
@@ -11,7 +11,8 @@ import 'package:shelf_static/shelf_static.dart';
 import '../../server.dart';
 import 'server_renderer.dart';
 
-final String kDevProxy = String.fromEnvironment('jaspr.dev.proxy', defaultValue: Platform.environment['jaspr_dev_proxy'] ?? '');
+final String kDevProxy =
+    String.fromEnvironment('jaspr.dev.proxy', defaultValue: Platform.environment['jaspr_dev_proxy'] ?? '');
 final String kDevFlutter = String.fromEnvironment('jaspr.dev.flutter');
 const bool kDevHotreload = bool.fromEnvironment('jaspr.dev.hotreload');
 
@@ -106,17 +107,17 @@ Handler _sseProxyHandler() {
   var serverUri = Uri.parse('http://localhost:$kDevProxy');
   var proxyPath = r'$dwdsSseHandler';
 
-  Handler? _incomingMessageProxyHandler;
-  var _httpClient = http.Client();
+  Handler? incomingMessageProxyHandler;
+  var httpClient = http.Client();
 
-  Future<Response> _createSseConnection(Request req, String path) async {
+  Future<Response> createSseConnection(Request req, String path) async {
     final serverReq = http.StreamedRequest(req.method, serverUri.replace(path: path, query: req.requestedUri.query))
       ..followRedirects = false
       ..headers.addAll(req.headers)
       ..headers['Host'] = serverUri.authority
       ..sink.close();
 
-    final serverResponse = await _httpClient.send(serverReq);
+    final serverResponse = await httpClient.send(serverReq);
 
     req.hijack((channel) {
       final sink = utf8.encoder.startChunkedConversion(channel.sink)..add(_sseHeaders(req.headers['origin']));
@@ -138,12 +139,12 @@ Handler _sseProxyHandler() {
     });
   }
 
-  Future<Response> _handleIncomingMessage(Request req, String path) async {
-    _incomingMessageProxyHandler ??= proxyHandler(
+  Future<Response> handleIncomingMessage(Request req, String path) async {
+    incomingMessageProxyHandler ??= proxyHandler(
       serverUri,
-      client: _httpClient,
+      client: httpClient,
     );
-    return _incomingMessageProxyHandler!(req);
+    return incomingMessageProxyHandler!(req);
   }
 
   return (Request req) async {
@@ -154,11 +155,11 @@ Handler _sseProxyHandler() {
     }
 
     if (req.headers['accept'] == 'text/event-stream' && req.method == 'GET') {
-      return _createSseConnection(req, path);
+      return createSseConnection(req, path);
     }
 
     if (req.headers['accept'] != 'text/event-stream' && req.method == 'POST') {
-      return _handleIncomingMessage(req, path);
+      return handleIncomingMessage(req, path);
     }
 
     return Response.notFound('');
@@ -182,9 +183,9 @@ final staticFileHandler = kDevProxy.isNotEmpty
     ? _proxyHandler()
     : createStaticHandler(join(projectDir, 'web'), defaultDocument: 'index.html');
 
-typedef _SetupHandler = FutureOr<Response> Function(Request, FutureOr<Response> Function(SetupFunction setup));
+typedef SetupHandler = FutureOr<Response> Function(Request, FutureOr<Response> Function(SetupFunction setup));
 
-Handler createHandler(_SetupHandler handle, {List<Middleware> middleware = const [], Handler? fileHandler}) {
+Handler createHandler(SetupHandler handle, {List<Middleware> middleware = const [], Handler? fileHandler}) {
   var staticHandler = fileHandler ?? staticFileHandler;
 
   var cascade = Cascade();
@@ -221,7 +222,6 @@ Handler createHandler(_SetupHandler handle, {List<Middleware> middleware = const
   for (var m in middleware) {
     pipeline = pipeline.addMiddleware(m);
   }
-  pipeline = pipeline.addMiddleware(logRequests());
 
   return RefreshableHandler(pipeline.addHandler(cascade.handler), onRefresh: () {
     if (staticHandler is RefreshableHandler) {