+++ packages/jaspr/CHANGELOG.md
@@ -1,4 +1,24 @@
-# 0.2.0
+## 0.4.0
+
+- **BREAKING** Bindings are no longer singletons. 
+  - `ComponentsBinding.instance`, `SchedulerBinding.instance` etc. were removed.
+  - You can access the current binding through `BuildContext`s `context.binding` property.
+
+- **BREAKING** Removed `ComponentTester.setUp()`, `BrowserTester.setUp()` and `ServerTester.setUp()`.
+  - Use `testComponents()`, `testBrowser()` and `testServer()` instead.
+
+- Requires Dart 3.0 or later.
+
+## 0.3.0
+
+- **BREAKING** The cli is now a separate package: `jaspr_cli`. To migrate run:
+  ```shell
+    dart pub global deactivate jaspr
+    dart pub global activate jaspr_cli
+  ```
+  The usage stays the same with `jaspr create`, `jaspr serve` and `jaspr build`.
+
+## 0.2.0
 
 **BREAKING**: This is the first major release after the initial publish and contains several breaking changes.
 
@@ -10,29 +30,29 @@
 - Added `Styles` class for writing typed css in dart.
 - Various other improvements throughout the framework.
 
-# 0.1.5
+## 0.1.5
 
 - Refactor cli code [#12](https://github.com/schultek/jaspr/pull/12)
 - Fixed issue with `jaspr serve`
 
-# 0.1.4
+## 0.1.4
 
 - Added `--port` option to `jaspr serve` ([#6](https://github.com/schultek/jaspr/issues/6))
 - Fixed `webdev` dependency constraint ([#5](https://github.com/schultek/jaspr/issues/5))
 
-# 0.1.3
+## 0.1.3
 
 - Fixed domino dependency ([#4](https://github.com/schultek/jaspr/pull/4))
 - Fixed build on windows ([#3](https://github.com/schultek/jaspr/issues/3))
 
-# 0.1.2
+## 0.1.2
 
 - Added executable to `pubspec.yaml` ([#2](https://github.com/schultek/jaspr/issues/2))
 
-# 0.1.1
+## 0.1.1
 
 - Fixed template for `jaspr create` ([#1](https://github.com/schultek/jaspr/pull/1))
 
-# 0.1.0
+## 0.1.0
 
 - Initial version