+++ packages/jaspr_riverpod/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.3.1
+
+- Added `ref.binding` to access the component binding inside providers.
+
+## 0.3.0
+
+- Update riverpod to `v2.3.2`
+
 ## 0.2.0
 
 - Update riverpod to `v2.1.3`